backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
psycopg[binary]==3.2.2
alembic==1.13.2
pydantic==2.8.2
pydantic-settings==2.4.0
email-validator==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
requests==2.32.3
redis==5.0.8
rq==1.16.2
Pillow==10.4.0
python-dateutil==2.9.0.post0